test moved to avoid visibility changes
