+++ json-path/src/test/java/com/jayway/jsonpath/internal/token/PathTokenTest.java
@@ -1,5 +1,6 @@
-package com.jayway.jsonpath;
+package com.jayway.jsonpath.internal.token;
 
+import com.jayway.jsonpath.BaseTest;
 import com.jayway.jsonpath.internal.token.PathToken;
 import com.jayway.jsonpath.internal.token.PropertyPathToken;
 import com.jayway.jsonpath.internal.token.ScanPathToken;
